Split encoder into multiple files. Unit tests and documentation.
