+++ test/protocol.spec.js
@@ -10,6 +10,7 @@ describe('When decoding collectd\'s binary protocol', function () {
     var defaultMock;
     var typesMock;
     var customTypeMock;
+    var customBinaryTypeMock;
 
     /*jshint -W117 */
     before(function () {
@@ -64,6 +65,20 @@ describe('When decoding collectd\'s binary protocol', function () {
             dstypes: [ 'gauge' ],
             values: [ 152567808.92 ],
             dsnames: [ 'value' ]
+        }];
+
+        customBinaryTypeMock = [{
+            host: 'localhost',
+            time: 1455098772,
+            interval: 10,
+            plugin: 'GenericJMX',
+            plugin_instance: 'MemoryPool|Eden_Space',
+            type: 'custom',
+            type_instance: 'committed',
+            dstypes: [ 'gauge' ],
+            values: [ 152567808.92 ],
+            dsnames: [ 'value' ],
+            tags: 'host=localhost,cluster=dev'
         }];
     });
 
@@ -324,32 +339,58 @@ describe('When decoding collectd\'s binary protocol', function () {
         });
     });
 
-    it('should decode metrics standard metrics with configured custom parts', function (done) {
+    it('should decode standard metrics metrics with configured custom parts', function (done) {
+        var binaryData = encoder.encodeCustom(customTypeMock, { 0x0099: 'tags' });
 
-        var customTypeMock = [{
-            host: 'localhost',
-            time: 1455098772,
-            interval: 10,
-            plugin: 'GenericJMX',
-            plugin_instance: 'MemoryPool|Eden_Space',
-            type: 'custom',
-            type_instance: 'committed',
-            dstypes: [ 'gauge' ],
-            values: [ 152567808.92 ],
-            dsnames: [ 'value' ],
-            tags: 'host=localhost,cluster=dev'
-        }];
-
-        var customPartConfig = { 0x0099: 'tags' };
-
-        var binaryData = encoder.encodeCustom(customTypeMock, customPartConfig);
-
-        var result = decoder.decodeCustom(binaryData, customPartConfig);
+        var result = decoder.decodeCustom(binaryData, { 0x0099: 'tags' });
 
         var decoded;
         result.on('data', function(data) {
             decoded = data;
-            console.log(decoded);
+        }).on('end', function () {
+            assert.deepEqual(decoded, customTypeMock);
+            done();
+        });
+    });
+
+    it('should not decode custom metrics with compliant decoder', function (done) {
+
+        var binaryData = encoder.encodeCustom(customBinaryTypeMock, { 0x0099: 'tags' });
+
+        var result = decoder.decode(binaryData);
+
+        var decoded;
+        result.on('data', function(data) {
+            decoded = data;
+        }).on('error', function (error) {
+            assert.throws(function () { throw error; }, Error, 'No handler for type 153');
+            done();
+        });
+    });
+
+    it('should decode custom metrics', function (done) {
+
+        var binaryData = encoder.encodeCustom(customBinaryTypeMock, { 0x0099: 'tags' });
+
+        var result = decoder.decodeCustom(binaryData, { 0x0099: 'tags' });
+
+        var decoded;
+        result.on('data', function(data) {
+            decoded = data;
+        }).on('end', function () {
+            assert.deepEqual(decoded, customBinaryTypeMock);
+            done();
+        });
+    });
+
+    it('should allow to instantiate encoder and decoder with invalid configuration', function () {
+        var binaryData = encoder.encodeCustom(customTypeMock, { 0x0001: 'tags' });
+
+        var result = decoder.decodeCustom(binaryData, { 0x0001: 'tags' });
+
+        var decoded;
+        result.on('data', function(data) {
+            decoded = data;
         }).on('end', function () {
             assert.deepEqual(decoded, customTypeMock);
             done();
